refa: refactor change menu domain and datasource folder
