+++ src/app/features/roles/domain/entities/index.ts
@@ -1,2 +1,3 @@
+export * from './menu-list.entity';
 export * from './roles-entity';
 export * from './roles-info-entity';
