+++ src/app/features/roles/domain/usecase/index.ts
@@ -1,5 +1,6 @@
 export * from './create-roles-usecase';
 export * from './delete-roles-usecase';
+export * from './get-menu-list.usecase';
 export * from './get-roles-info-usecase';
 export * from './get-roles-usecase';
 export * from './update-roles-usecase';
