+++ src/app/features/roles/infrastructure/dtos/responses/index.ts
@@ -1,1 +1,2 @@
+export * from './menu-list.dto';
 export * from './roles-dto-response';
