+++ src/app/features/roles/infrastructure/repositories/roles-repository-impl.ts
@@ -4,7 +4,8 @@ import { HttpResourceRef } from "@angular/common/http";
 
 import { ApiResponse } from "@utils/api_response";
 import { RolesDatasource } from "@app/roles/infrastructure/datasource";
-import { RolesEntity, RolesInfoEntity, RolesRepository } from "@app/roles/domain";
+import { MenuRolesMapper } from "@app/roles/infrastructure/mappers/menu-list.mapper";
+import { MenuListResponseEntity, RolesEntity, RolesInfoEntity, RolesRepository } from "@app/roles/domain";
 
 @Injectable({ providedIn: 'root'})
 export class RolesRepositoryImpl implements RolesRepository {
@@ -21,6 +22,12 @@ export class RolesRepositoryImpl implements RolesRepository {
 
     getAllRoles(limit: number, offset: number): Promise<ApiResponse<RolesEntity[]>> {
         return this.datasource.getAllRoles( limit, offset );
+    }
+
+    async getMenuList( limit = 5, offset = 0 ): Promise<ApiResponse<MenuListResponseEntity[]>> {
+        const { data, message, status } = await this.datasource.getMenuList( limit, offset );
+        const dataMapped = data.map( userMenuRol => MenuRolesMapper.fromResponseDto(userMenuRol) );
+        return { message, status, data: dataMapped } as ApiResponse<MenuListResponseEntity[]>;
     }
 
     createNewRol(newRol: string): Observable<ApiResponse<RolesEntity>> {
