+++ src/app/features/users/domain/entities/index.ts
@@ -1,3 +1,2 @@
-export * from './menu-roles.entity';
 export * from './users-menu-roles.entity';
 export * from './create-menu-roles.entity';
