--- src/app/features/users/domain/entities/menu-roles.entity.ts
@@ -1,7 +1,0 @@
-export interface MenuListResponseEntity {
-    idMenu: number;
-    name:   string;
-    route:  string;
-    icon:   string;
-    active: boolean;
-}
