--- src/app/features/users/domain/usecase/get-menu-list.usecase.ts
@@ -1,8 +1,0 @@
-import { Injectable } from "@angular/core";
-import { UsersMenuRolesRepository } from "@app/users/domain/repositories";
-
-@Injectable({providedIn: 'root'})
-export class GetMenuListUsecase {
-    constructor( private repository:UsersMenuRolesRepository ) {}
-    execute(limit: number, offset: number) { return this.repository.getMenuList(); }
-}
