+++ src/app/features/users/domain/usecase/index.ts
@@ -1,5 +1,4 @@
 export * from './delete-user.usecase';
-export * from './get-menu-list.usecase';
 export * from './get-user-menu-list.usecase';
 export * from './update-user-menu-list.usecase';
 export * from './create-user-menu-list.usecase';
