+++ src/app/features/users/infrastructure/dtos/responses/index.ts
@@ -1,2 +1,1 @@
-export * from './menu-list.dto';
 export * from './users-menu-roles.dto';
