--- src/app/features/users/infrastructure/dtos/responses/menu-list.dto.ts
@@ -1,7 +1,0 @@
-export interface MenuListResponseDto {
-    idMenu: number;
-    name:   string;
-    route:  string;
-    icon:   string;
-    active: boolean;
-}
