+++ src/app/features/users/infrastructure/mappers/index.ts
@@ -1,3 +1,2 @@
-export * from './menu-list.mapper';
 export * from './user-menu-roles.mapper';
 export * from './create-menu-roles.mapper';
