--- src/app/features/users/infrastructure/mappers/menu-list.mapper.ts
@@ -1,8 +1,0 @@
-import { MenuListResponseEntity} from "@app/users/domain/entities";
-import { MenuListResponseDto } from "@app/users/infrastructure/dtos/responses";
-
-export class MenuRolesMapper {
-    static fromResponseDto( dto:MenuListResponseDto ):MenuListResponseEntity {
-        return { ...dto };
-    }
-}
