+++ src/app/features/users/infrastructure/repositories/users-menu-roles.repository.impl.ts
@@ -3,8 +3,8 @@ import { Injectable } from "@angular/core";
 import { ApiResponse } from "@utils/api_response";
 import { UsersMenuRolesRepository } from "@app/users/domain/repositories";
 import { UserMenuRolesDatasource } from "@app/users/infrastructure/datasource";
-import { CreateUserMenuRolesMapper, MenuRolesMapper, UserMenuRolesMapper } from "@app/users/infrastructure/mappers";
-import { UsersMenuRolesEntity, MenuListResponseEntity, CreateUserMenuRolesEntity } from "@app/users/domain/entities";
+import { UsersMenuRolesEntity, CreateUserMenuRolesEntity } from "@app/users/domain/entities";
+import { CreateUserMenuRolesMapper, UserMenuRolesMapper } from "@app/users/infrastructure/mappers";
 
 @Injectable({ providedIn: 'root'})
 export class UsersMenuRolesRepositoryImpl implements UsersMenuRolesRepository {
@@ -15,12 +15,6 @@ export class UsersMenuRolesRepositoryImpl implements UsersMenuRolesRepository {
         const { data, message, status } = await this.datasource.getUserMenuList( limit, offset );
         const dataMapped = data.map( userMenuRol => UserMenuRolesMapper.fromResponseDto(userMenuRol) );
         return { message, status, data: dataMapped } as ApiResponse<UsersMenuRolesEntity[]>;
-    }
-
-    async getMenuList( limit = 5, offset = 0 ): Promise<ApiResponse<MenuListResponseEntity[]>> {
-        const { data, message, status } = await this.datasource.getMenuList( limit, offset );
-        const dataMapped = data.map( userMenuRol => MenuRolesMapper.fromResponseDto(userMenuRol) );
-        return { message, status, data: dataMapped } as ApiResponse<MenuListResponseEntity[]>;
     }
 
     async createUserMenuRoles( createMenuUserRole:CreateUserMenuRolesEntity ) : Promise<ApiResponse<any[]>> {
